% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{BinaryContiInC}
\alias{BinaryContiInC}
\title{Estimate the correlation for binary and continuous data based on Kendall's rank cor}
\usage{
BinaryContiInC(x, y)
}
\arguments{
\item{x}{A vector of binary variables}

\item{y}{A vector of continuous variables}
}
\description{
Estimate the correlation for binary and continuous data based on Kendall's rank cor
}
