% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{BinaryInC}
\alias{BinaryInC}
\title{Estimate the correlation for binary data based on Kendall's rank cor}
\usage{
BinaryInC(x, y)
}
\arguments{
\item{x}{A vector of binary variables}

\item{y}{A vector of binary variables}
}
\description{
Estimate the correlation for binary data based on Kendall's rank cor
}
