% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{CategoricalCategorical}
\alias{CategoricalCategorical}
\title{Estimate the correlation for categorical data based on Kendall's rank cor}
\usage{
CategoricalCategorical(x, y)
}
\arguments{
\item{x}{A vector of categorical variables}

\item{y}{A vector of categorical variables}
}
\description{
Estimate the correlation for categorical data based on Kendall's rank cor
}
