% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{ContinuousInC}
\alias{ContinuousInC}
\title{Estimate the correlation for continuous data based on Kendall's rank cor}
\usage{
ContinuousInC(x, y)
}
\arguments{
\item{x}{A vector of continuous variables}

\item{y}{A vector of continuous variables}
}
\description{
Estimate the correlation for continuous data based on Kendall's rank cor
}
