% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{kendall_tau}
\alias{kendall_tau}
\title{Compute the Kendall' tau rank correlation}
\usage{
kendall_tau(x, y)
}
\arguments{
\item{x}{A vector}

\item{y}{A vector}
}
\description{
Compute the Kendall' tau rank correlation
}
