% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{LGC_sigma_AllType}
\alias{LGC_sigma_AllType}
\title{Compute the latent covariance matrix when data is all type}
\usage{
LGC_sigma_AllType(x, label)
}
\arguments{
\item{x}{A single matrix}

\item{label}{A vector of labels}
}
\description{
Compute the latent covariance matrix when data is all type
}
