% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{LGC_sigma_BC}
\alias{LGC_sigma_BC}
\title{Compute the latent covariance matrix when data is binary and continuous}
\usage{
LGC_sigma_BC(x, label)
}
\arguments{
\item{x}{A single matrix}

\item{label}{A vector of labels}
}
\description{
Compute the latent covariance matrix when data is binary and continuous
}
