% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{OrdinalOrdinal}
\alias{OrdinalOrdinal}
\title{Estimate the correlation for ordinal data based on Kendall's rank cor}
\usage{
OrdinalOrdinal(x, y)
}
\arguments{
\item{x}{A vector of ordinal variables}

\item{y}{A vector of ordinal variables}
}
\description{
Estimate the correlation for ordinal data based on Kendall's rank cor
}
